remove methods already tested by unit test from integration test
