+++ streams/src/test/java/org/apache/kafka/streams/tools/StreamsResetterTest.java
@@ -1,4 +1,20 @@
-package unit.kafka.tools;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.kafka.streams.tools;
 
 import kafka.tools.StreamsResetter;
 import org.apache.kafka.clients.NodeApiVersions;
@@ -33,7 +49,7 @@ import static org.junit.Assert.fail;
  */
 public class StreamsResetterTest {
 
-    public static final String TOPIC = "topic1";
+    private static final String TOPIC = "topic1";
     private final StreamsResetter streamsResetter = new StreamsResetter();
     private final MockConsumer<byte[], byte[]> consumer = new MockConsumer<>(OffsetResetStrategy.EARLIEST);
     private final TopicPartition topicPartition = new TopicPartition(TOPIC, 0);
